refactor(node): clean up debuglog
